bugfeed/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-guardian==2.2.0
django-oauth-toolkit==1.3.2
django-richtextfield==1.5.0
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.1
pkg-resources==0.0.0
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
